requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
execnet==2.1.1
Faker==37.12.0
greenlet==3.2.4
idna==3.10
iniconfig==2.0.0
mysql-connector-python==9.5.0
packaging==24.2
playwright==1.55.0
pluggy==1.5.0
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
pytest==8.3.4
pytest-base-url==2.1.0
pytest-playwright==0.7.0
pytest-xdist==3.6.1
python-slugify==8.0.4
python-stdnum==2.1
requests==2.32.3
tenacity==9.1.2
text-unidecode==1.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.3.0
